+++ src/Route.tsx
@@ -1,5 +1,6 @@
-import React from "react";
+import React, { useMemo } from "react";
 
+import { useGlobalContext } from "./context/GlobalContext";
 import { RequestResponse } from "./common/route/RequestResponse";
 import { Documentation } from "./common/route/Documentation";
 
@@ -11,6 +12,7 @@ import {
   WorkspaceConfig,
   AuthForm,
   HistoryMetadata,
+  HistoricResponse,
 } from "./types";
 
 export default function Route({
@@ -26,16 +28,32 @@ export default function Route({
   HistoryMetadata?: HistoryMetadata;
   workspaceConfig: WorkspaceConfig;
 }) {
+  const { historicResponses } = useGlobalContext();
+
+  // If `activeRoute` is specified, filter displayed History to records matching just that route
+  const filteredHistory = useMemo(() => {
+    return !route
+      ? historicResponses
+      : historicResponses.filter(
+          (historicResponse: HistoricResponse) =>
+            historicResponse?.route?.path === route.path
+        );
+  }, [historicResponses, route]);
+
   return (
     <>
       {AuthForm && workspaceConfig ? (
         <AuthForm workspaceConfig={workspaceConfig} />
       ) : null}
       <RequestResponse
+        filteredHistory={filteredHistory}
         route={route}
         applyAxiosInterceptors={applyAxiosInterceptors}
       />
-      <History activeRoute={route} HistoryMetadata={HistoryMetadata} />
+      <History
+        filteredHistory={filteredHistory}
+        HistoryMetadata={HistoryMetadata}
+      />
       <Documentation documentation={route.documentation} />
     </>
   );