+++ src/context/GlobalContext.tsx
@@ -4,19 +4,44 @@ import * as storage from "../lib/storage";
 
 const storageKeys = {
   darkMode: "darkMode",
+  hideDeprecatedRoutes: "hideDeprecatedRoutes",
   historicResponses: "historic-responses",
+  collapsedWorkspaces: "collapsed-workspaces",
 };
 
-import { HistoricResponse, StoreHistoricResponsePayload } from "../types";
+import { HistoricResponse, Route } from "../types";
 
 export const Context = React.createContext({
   darkMode: storage.get(storageKeys.darkMode),
   setDarkMode: (darkMode: boolean) => {
     // noop
   },
+  hideDeprecatedRoutes: storage.get(storageKeys.hideDeprecatedRoutes),
+  setHideDeprecatedRoutes: (hideDeprecatedRoutes: boolean) => {
+    // noop
+  },
   historicResponses: (storage.get(storageKeys.historicResponses) ||
     []) as HistoricResponse[],
-  storeHistoricResponse: (historicResponse: StoreHistoricResponsePayload) => {
+  storeHistoricResponse: (historicResponse: HistoricResponse) => {
+    return historicResponse;
+  },
+  partialRequestResponses: {} as Record<string, HistoricResponse>,
+  setPartialRequestResponse: (historicResponse: HistoricResponse) => {
+    // noop
+  },
+
+  activeRoute: null as null | Route,
+  setActiveRoute: (activeRoute: Route) => {
+    // noop
+  },
+
+  keywords: "",
+  setKeywords: (keywords: string) => {
+    // noop
+  },
+
+  collapsedWorkspaces: [] as string[],
+  setCollapsedWorkspaces: (collapsedWorkspaces: string[]) => {
     // noop
   },
 });
@@ -24,14 +49,56 @@ export const Context = React.createContext({
 export const useGlobalContext = () => useContext(Context);
 
 export function ContextProvider({ children }: { children: React.ReactNode }) {
+  const [activeRoute, setActiveRoute] = useState<null | Route>(null);
+  const [keywords, setKeywords] = useState("");
+  const [collapsedWorkspaces, setCollapsedWorkspacesInState] = useState<
+    string[]
+  >(storage.get(storageKeys.collapsedWorkspaces) || []);
   const [darkMode, setDarkModeInState] = useState<boolean>(
     storage.get(storageKeys.darkMode)
+  );
+
+  // Used to track the state of a Request for a particular Route before it becomes a HistoricResponse
+  // after the Request is issued. This allows the user to flip back and forth between routes and still have
+  // Params stay loaded.
+  const [partialRequestResponses, setPartialRequestResponses] = useState<
+    Record<string, HistoricResponse>
+  >({});
+
+  const setPartialRequestResponse = useCallback(
+    (partialRequestResponse: HistoricResponse) => {
+      setPartialRequestResponses({
+        ...partialRequestResponses,
+        [partialRequestResponse.route.path]: partialRequestResponse,
+      });
+    },
+    [partialRequestResponses]
   );
 
   const setDarkMode = useCallback((darkMode: boolean) => {
     storage.set(storageKeys.darkMode, darkMode);
     setDarkModeInState(darkMode);
   }, []);
+
+  const [hideDeprecatedRoutes, setHideDeprecatedRoutesInState] = useState<
+    boolean
+  >(storage.get(storageKeys.hideDeprecatedRoutes));
+
+  const setCollapsedWorkspaces = useCallback(
+    (collapsedWorkspaces: string[]) => {
+      storage.set(storageKeys.collapsedWorkspaces, collapsedWorkspaces);
+      setCollapsedWorkspacesInState(collapsedWorkspaces);
+    },
+    []
+  );
+
+  const setHideDeprecatedRoutes = useCallback(
+    (hideDeprecatedRoutes: boolean) => {
+      storage.set(storageKeys.hideDeprecatedRoutes, hideDeprecatedRoutes);
+      setHideDeprecatedRoutesInState(hideDeprecatedRoutes);
+    },
+    []
+  );
 
   const [historicResponses, setHistoricResponseInState] = useState<
     HistoricResponse[]
@@ -57,7 +124,7 @@ export function ContextProvider({ children }: { children: React.ReactNode }) {
       qsParams,
       body,
       urlParams,
-    }: StoreHistoricResponsePayload) => {
+    }: HistoricResponse) => {
       const newResponse: HistoricResponse = {
         route,
         qsParams,
@@ -92,11 +159,13 @@ export function ContextProvider({ children }: { children: React.ReactNode }) {
 
       const newHistoricResponses = [newResponse, ...historicResponses].slice(
         0,
-        50
-      ); // prepend the new HistoricResponse, and ensure the array has a max of 50 cells
+        100
+      ); // prepend the new HistoricResponse, and ensure the array has a max of 100 cells
 
       storage.set(storageKeys.historicResponses, newHistoricResponses);
       setHistoricResponseInState(newHistoricResponses);
+
+      return newResponse;
     },
     [historicResponses]
   );
@@ -106,8 +175,18 @@ export function ContextProvider({ children }: { children: React.ReactNode }) {
       value={{
         darkMode,
         setDarkMode,
+        hideDeprecatedRoutes,
+        setHideDeprecatedRoutes,
         historicResponses,
         storeHistoricResponse,
+        partialRequestResponses,
+        setPartialRequestResponse,
+        activeRoute,
+        setActiveRoute,
+        keywords,
+        setKeywords,
+        collapsedWorkspaces,
+        setCollapsedWorkspaces,
       }}
     >
       {children}